move MatchFeatureData class to vespajlib
