+++ vespajlib/src/main/java/com/yahoo/data/access/helpers/package-info.java
@@ -1,0 +1,5 @@
+// Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+@ExportPackage
+package com.yahoo.data.access.helpers;
+
+import com.yahoo.osgi.annotation.ExportPackage;
