+++ vespajlib/src/test/java/com/yahoo/data/access/helpers/MatchFeatureDataTest.java
@@ -1,6 +1,6 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-package com.yahoo.search.dispatch.rpc;
+package com.yahoo.data.access.helpers;
 
 import com.yahoo.data.access.ArrayTraverser;
 import com.yahoo.data.access.Inspector;
